+++ doc-utils/ComponentExample.tsx
@@ -5,13 +5,14 @@ import { ExampleBase } from "./ExampleBase";
 export interface Props {
     component: FC<ExampleComponentProps>;
     title?: string;
+    scrollable?: boolean;
     knobs?: {
         boolProps?: Array<BoolProp>;
         choiceProps?: Array<ChoiceProp>;
     };
 }
 
-export const ComponentExample: FC<Props> = ({ component, knobs, title = "Komponent", children }) => {
+export const ComponentExample: FC<Props> = ({ component, knobs, title = "Komponent", children, ...rest }) => {
     const C = component;
 
     return (
@@ -20,6 +21,7 @@ export const ComponentExample: FC<Props> = ({ component, knobs, title = "Kompone
             title={title}
             responsiveLayout={false}
             component={({ boolValues, choiceValues }) => <C boolValues={boolValues} choiceValues={choiceValues} />}
+            {...rest}
         >
             {children}
         </ExampleBase>