+++ doc-utils/ExampleBase.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useLayoutEffect, FC } from "react";
-import classNames from "classnames";
+import cx from "classnames";
 import { nanoid } from "nanoid";
 import { Checkbox } from "@fremtind/jkl-checkbox-react";
 import { RadioButtons } from "@fremtind/jkl-radio-button-react";
@@ -10,6 +10,7 @@ import { hyphenate } from "./internal/hypenate";
 export interface Props {
     component: FC<ExampleComponentProps>;
     title?: string;
+    scrollable?: boolean;
     knobs?: {
         boolProps?: Array<BoolProp>;
         choiceProps?: Array<ChoiceProp>;
@@ -17,7 +18,14 @@ export interface Props {
     responsiveLayout: boolean;
 }
 
-export const ExampleBase: FC<Props> = ({ responsiveLayout, component, knobs, title = "Komponent", children }) => {
+export const ExampleBase: FC<Props> = ({
+    responsiveLayout,
+    component,
+    knobs,
+    title = "Komponent",
+    children,
+    scrollable,
+}) => {
     const [uid] = useState(`example-${nanoid(8)}`);
     const [showCode, setShowCode] = useState(false);
     const [boolValues, setBoolValues] = useState<Dictionary<boolean>>({});
@@ -53,13 +61,14 @@ export const ExampleBase: FC<Props> = ({ responsiveLayout, component, knobs, tit
     const C = component;
 
     return (
-        <div className="jkl-layout-spacing--large-bottom">
+        <div className="jkl-spacing-2xl--bottom">
             <section className={responsiveLayout ? "jkl-portal-responsive-example" : "jkl-portal-component-example"}>
                 <div
                     data-theme={theme}
                     data-example-text={title}
-                    className={classNames("jkl-portal-component-example__example-wrapper", {
+                    className={cx("jkl-portal-component-example__example-wrapper", {
                         "jkl-portal-component-example__example-wrapper--dark": theme === "dark",
+                        "jkl-portal-component-example__example-wrapper--scrollable": scrollable,
                     })}
                 >
                     <C boolValues={boolValues} choiceValues={choiceValues} />
@@ -118,14 +127,11 @@ export const ExampleBase: FC<Props> = ({ responsiveLayout, component, knobs, tit
             </section>
             {children && (
                 <div
-                    className={classNames("jkl-portal-code-example", {
+                    className={cx("jkl-portal-code-example", {
                         "jkl-portal-code-example--open": !showCode,
                     })}
                 >
-                    <button
-                        className={"jkl-button jkl-button--tertiary jkl-component-spacing--small-top"}
-                        onClick={toggleCode}
-                    >
+                    <button className={"jkl-button jkl-button--tertiary jkl-spacing-xs--top"} onClick={toggleCode}>
                         {`${showCode ? "Skjul" : "Vis"} kode`}
                     </button>
                     {showCode && children}