+++ portal/src/components/Experimental/Experimental.tsx
@@ -3,10 +3,7 @@ import { WarningMessageBox } from "@fremtind/jkl-message-box-react";
 
 export const Experimental: React.FC = ({ children }) => {
     return (
-        <WarningMessageBox
-            className="jkl-layout-spacing--medium-top jkl-layout-spacing--medium-bottom"
-            title="Eksperimentell komponent"
-        >
+        <WarningMessageBox className="jkl-spacing-xl--top jkl-spacing-xl--bottom" title="Eksperimentell komponent">
             {children}
         </WarningMessageBox>
     );