+++ portal/src/components/Layout/DocPageLayout.tsx
@@ -1,7 +1,7 @@
 import React, { FC, ReactNode, useContext } from "react";
 import { motion } from "framer-motion";
 import { Accordion, AccordionItem } from "@fremtind/jkl-accordion-react";
-import { Table } from "@fremtind/jkl-table-react";
+import { DataTable } from "@fremtind/jkl-table-react";
 import { Frontmatter, FrontmatterTypeProp } from "../Header/useNavigationLinks";
 import { BlogPageHeader, ComponentPageHeader } from "./components";
 import { a11yContext } from "../../contexts/a11yContext";
@@ -42,14 +42,15 @@ export const DocPageLayout: FC<Props> = ({ children, location, pageContext: { fr
             <BlogPageHeader {...frontmatter} />
             {children}
             {frontmatter.type && (
-                <section className="jkl-layout-spacing--xl-bottom jkl-portal-paragraph">
-                    <h2 className="jkl-heading-3 jkl-layout-spacing--xl-top">PropTypes</h2>
-                    <Accordion className="jkl-layout-spacing--medium-top">
+                <section className="jkl-spacing-3xl--bottom jkl-portal-paragraph">
+                    <h2 className="jkl-heading-3 jkl-spacing-3xl--top">PropTypes</h2>
+                    <Accordion className="jkl-spacing-xl--top">
                         {frontmatter.type.map(({ displayName, props }) => (
                             <AccordionItem title={`${displayName}`} key={displayName}>
-                                <Table
+                                <DataTable
                                     columns={["Prop", "DefaultValue", "Required", "Type"]}
                                     rows={getRows(Object.values(props))}
+                                    verticalAlign="top"
                                 />
                             </AccordionItem>
                         ))}