+++ packages/message-box-react/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 5.0.1 (2021-10-25)
+
+### Bug Fixes
+
+- enabled other components as children to messagebox ([#2407](https://github.com/fremtind/jokul/issues/2407)) ([f531c68](https://github.com/fremtind/jokul/commit/f531c68e88eb01cceecfb6a0133efc4efd66df99))
+
+## 5.0.0 (2021-10-21)
+
+### chore
+
+- updates dependencies ([b975200](https://github.com/fremtind/jokul/commit/b97520045c02e4bcb44ebde159c60a7dff7f01d6))
+
+### BREAKING CHANGES
+
+- depends on jkl-core v8 with breaking changes
+
 ## 4.2.3 (2021-10-20)
 
 ### Bug Fixes